+++ MochiKit/LoggingPane.js
@@ -77,7 +77,6 @@ MochiKit.LoggingPane.LoggingPane = function (inline/* = false */, logger/* = Moc
             alert("Not able to open debugging window due to pop-up blocking.");
             return undefined;
         }
-        nwin.document.open();
         nwin.document.write(
             '<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" '
             + '"http://www.w3.org/TR/html4/loose.dtd">'
@@ -134,8 +133,8 @@ MochiKit.LoggingPane.LoggingPane = function (inline/* = false */, logger/* = Moc
         var text = messageText(msg);
         var c = this.colorTable[level];
         var p = doc.createElement("span");
-        p.style.color = c;
-        p.style.margin = "0";
+        p.className = "MochiKit-LogMessage MochiKit-LogLevel-" + level;
+        p.style.cssText = "margin: 0px; white-space: -moz-pre-wrap; white-space: -o-pre-wrap; white-space: pre-wrap; white-space: pre-line; word-wrap: break-word; wrap-option: emergency; color: " + c;
         p.appendChild(doc.createTextNode(level + ": " + text));
         logPane.appendChild(p);
         logPane.appendChild(doc.createElement("br"));
@@ -236,26 +235,18 @@ MochiKit.LoggingPane.LoggingPane = function (inline/* = false */, logger/* = Moc
     }, this);
 
     /* Create the debug pane */
-    var style = {
-        "display": "block",
-        "left": "0px",
-        "bottom": "0px",
-        "position": "fixed",
-        "font": this.logFont,
-        "width": "100%",
-        "height": "100%",
-        "backgroundColor": "white"
-    };
+    var style = "display: block; left: 0px; bottom: 0px; position: fixed; width: 100%; background-color: white; font: " + this.logFont;
     if (inline) {
-        style.height = "10em";
-        style.borderTop = "2px solid black";
+        style += "; height: 10em; border-top: 2px solid black";
+    } else {
+        style += "; height: 100%;";
     }
-    update(debugPane.style, style);
+    debugPane.style.cssText = style;
 
     doc.body.appendChild(debugPane);
 
     /* Create the filter fields */
-    style = {"width": "33%", "display": "inline", "font": this.logFont};
+    style = {"cssText": "width: 33%; display: inline; font: " + this.logFont};
 
     updatetree(levelFilterField, {
         "value": "FATAL|ERROR|WARNING|INFO|DEBUG",
@@ -272,40 +263,31 @@ MochiKit.LoggingPane.LoggingPane = function (inline/* = false */, logger/* = Moc
     debugPane.appendChild(infoFilterField);
 
     /* Create the buttons */
-    style = {"width": "8%", "display": "inline", "font": this.logFont};
+    style = "width: 8%; display:inline; font: " + this.logFont;
 
     filterButton.appendChild(doc.createTextNode("Filter"));
     filterButton.onclick = bind(this.buildAndApplyFilter, this);
-    update(filterButton.style, style);
+    filterButton.style.cssText = style;
     debugPane.appendChild(filterButton);
 
     loadButton.appendChild(doc.createTextNode("Load"));
     loadButton.onclick = loadMessages;
-    update(loadButton.style, style);
+    loadButton.style.cssText = style;
     debugPane.appendChild(loadButton);
 
     clearButton.appendChild(doc.createTextNode("Clear"));
     clearButton.onclick = clearMessages;
-    update(clearButton.style, style);
+    clearButton.style.cssText = style;
     debugPane.appendChild(clearButton);
 
     closeButton.appendChild(doc.createTextNode("Close"));
     closeButton.onclick = closePane;
-    update(closeButton.style, style);
+    closeButton.style.cssText = style;
     debugPane.appendChild(closeButton);
 
     /* Create the logging pane */
-    debugPane.style.display = "block";
-    logPaneArea.style.overflow = "auto";
-    logPaneArea.style.width = "100%";
-    logPane.style.whiteSpace = "pre";
-    logPane.style.wordWrap = "break-word";
-    logPane.style.width = "100%";
-    if (inline) {
-        logPane.style.height = "8em";
-    } else {
-        logPane.style.height = "100%";
-    }
+    logPaneArea.style.cssText = "overflow: auto; width: 100%";
+    logPane.style.cssText = "width: 100%; height: " + (inline ? "8em" : "100%");
 
     logPaneArea.appendChild(logPane);
     debugPane.appendChild(logPaneArea);